#388 - Move StopwordRemover into tokit module
- Moved the code and the unit test
