+++ dkpro-core-tokit-asl/src/main/java/org/dkpro/core/tokit/internal/StopWordSet.java
@@ -15,7 +15,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.dkpro.core.stopwordremover;
+package org.dkpro.core.tokit.internal;
 
 import java.io.BufferedReader;
 import java.io.FileReader;
