+++ dkpro-core-tokit-asl/src/test/java/org/dkpro/core/tokit/StopWordRemoverTest.java
@@ -15,7 +15,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.dkpro.core.stopwordremover;
+package org.dkpro.core.tokit;
 
 import static org.apache.uima.fit.factory.AnalysisEngineFactory.createEngineDescription;
 import static org.apache.uima.fit.util.JCasUtil.select;
@@ -26,6 +26,7 @@ import org.apache.uima.analysis_engine.AnalysisEngineDescription;
 import org.apache.uima.jcas.JCas;
 import org.dkpro.core.testing.DkproTestContext;
 import org.dkpro.core.testing.TestRunner;
+import org.dkpro.core.tokit.StopWordRemover;
 import org.junit.Rule;
 import org.junit.Test;
