+++ dkpro-core-tokit-asl/src/test/resources/stopwords1.txt
